Fix: IOS-1866 Section front nodes sometimes stop responding to taps
The views backing ASDK nodes get removed from the section front collection
views on reload. If a touch is active on a view when it's removed from the
view hierarchy, the system continues tracking it, but the responder chain
is broken so touchesEnded: and touchesCancelled: aren't sent. UICollectionView
relies on getting one of these messages to clear the UITouch it started
tracking in touchesBegan:.

This patch shadows all of the touch methods on ASCollectionView, keeping
track of the touches ourself. If we receive touchesBegan: while we're still
tracking other touches, we send touchesCancelled: to the UICollectionView
to clear this state and avoid an unresponsive situation.